moved more code into the 'generic' module
