+++ Rust/bridges/dynamic_hash/src/interop.rs
@@ -11,7 +11,7 @@ use std::{
     sync::{Once, OnceLock},
 };
 
-use hashcat_sys::{ThreadContext, bridge_context_t, generic::string_from_ptr, generic_io_tmp_t};
+use hashcat_sys::generic::{ThreadContext, bridge_context_t, string_from_ptr, generic_io_tmp_t};
 
 use crate::{Expr, eval::EvalContext, parse};
