+++ Rust/bridges/generic_hash/src/generic_hash.rs
@@ -2,7 +2,7 @@
  * Author......: See docs/credits.txt
  * License.....: MIT
  */
-use hashcat_sys::ThreadContext;
+use hashcat_sys::generic::ThreadContext;
 use sha2::{Digest, Sha256};
 
 // Trailing zeroes are necessary.
